requirements.txt
# -----------------------------
# Core Backend
# -----------------------------
fastapi>=0.110
uvicorn[standard]>=0.29

# -----------------------------
# Database (SQLAlchemy)
# -----------------------------
sqlalchemy>=2.0
psycopg2-binary>=2.9        # PostgreSQL (sync)

# -----------------------------
# ORM & Data Validation
# -----------------------------
pydantic>=2.6
python-dotenv>=1.0
pydantic[email]>=2.6

# -----------------------------
# Authentication & Security
# -----------------------------
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
       # JWT token creation/verification

# -----------------------------
# LangChain + Groq
# -----------------------------
langchain>=0.1.17
langchain-community>=0.0.30
langchain-groq>=0.1.3
langchain-prompty>=1.0.0

# -----------------------------
# HTTP & Async Utilities
# -----------------------------
httpx>=0.27

# -----------------------------
# Utilities
# -----------------------------
orjson>=3.10
pdfplumber
python-multipart==0.0.6
cloudinary == 1.44.1